requirements.txt
smartapi-python
pyotp
anthropic
python-dotenv
logzero
requests
websocket
google-genai
